feat: controlplane gateway API + CLI, part 0 - skeleton (#17)
This PR brings the skeleton of the control plane and tries to tie together the CLI, control plane, shared data structures and the data plane.
There are three parts:

    Controlplane with gRPC gateway API.
    Golang bindings to the API of shared-memory structures (currently not implemented).
    Modular CLI.

The first one - the controlplane initializes its gRPC gateway, runs built-in modules, which we decided to put into a single binary, waits them to be registered and then just proxies all requests to them.
Also there are several utility services, such as logging, that allows to dynamically change logging level.

Proxying is implemented without gRPC reflection, but this can be changed later. For now, we just proxy each request to a module to the module itself, but we will change this in the near future by implementing a transactional API.

CLI is implemented as a separate set of binaries united under the core one, which also proxies command-line arguments to a specific one.